Change DC-DC layout
The ripple on the 5V and 3V3 rail was too high (360mV).

In an attempt to reduce the RF and HF noise, I changed the layout
to a more compact one. I also added a third 22uF output capacitor
behind the inductor to lower ESR. I also added a 100nf input
capacitor to combat HF noise, as recommended in the datasheet.

The BOM needs to be updated in such a way, that we specificly
choose low ESR resistors.